requirements.txt

Pillow==8.4.0
psutil==5.8.0
pywin32==302
pyinstaller==4.7
pyperclip==1.8.2
pytest==6.2.5